Fix empty details
Add a specific check for an empty text editor and notify the callback by
passing null. This is because even if the editor is empty the HTML output won't
be, instead it will be an empty p tag "<p></p>".

Note that we have hacky empty string setting because of the restricting nature
of the current API system. All changes are model updates, where a value of
undefined means "don't update the field" and a value of null means "delete the
field". However, in the current system null values are ignored by the server
for some reason, so there is no actual way to delete a field. Instead, we have
to resort to setting an empty string which evaluates to false in JS truthiness
checks (which nearly all checks are, but still hacky to rely on that...)