Fix location chooser in browser history
Pushing the rewrite URL (/opportunities) directly causes this error when we
go back:
https://nextjs.org/docs/messages/incompatible-href-as

In order to mitigate this we directly push the default location but rewrite it
as /opportunities. For some reason this causes a refresh, but it is the best
compromise I can come up with since it only happens when the user goes back to
the location selector after having already chosen a location (which should
be rare).